🚚 Move protected route components in `ConditionalRoute` component folder
Since a previous commit, `AuthenticatedRoute`, `GuestRoute` et `AdminRoute` rely on the `ConditionalRoute` component as they share the same logic. Moving their files to the `ConditionalRoute` directory improves organization by emphasizing this shared purpose.
