+++ src/components/Router/ConditionalRoute/AdminRoute.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import { useDispatch, useSelector } from 'react-redux'
 import { getIsAdmin } from '../../../redux/selectors/user';
 import { ErrorCode, setErrorPage } from '../../../redux/reducers/errorPage';
-import ConditionalRoute from '../ConditionalRoute';
+import ConditionalRoute from '.';
 
 export default function AdminRoute({ redirectTo, children }) {
     const dispatch = useDispatch()
