+++ src/components/Router/ConditionalRoute/GuestRoute.jsx
@@ -2,7 +2,7 @@ import PropTypes from 'prop-types'
 import { useDispatch, useSelector } from 'react-redux'
 import { getIsLogged } from '../../../redux/selectors/user'
 import { ErrorCode, setErrorPage } from '../../../redux/reducers/errorPage'
-import ConditionalRoute from '../ConditionalRoute'
+import ConditionalRoute from '.'
 
 export default function GuestRoute({ redirectTo, children }) {
     const dispatch = useDispatch()
